backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
torch==2.6.0
torchvision==0.21.0
pillow==12.1.1
numpy==1.26.4
opencv-python==4.11.0.86
python-multipart==0.0.22
